agent/requirements.txt
langchain_chroma==0.2.4
langchain_community==0.3.27
langchain_core==0.3.68
langchain_openai==0.3.27
langchain_text_splitters==0.3.8
python-dotenv